Rename NewtonianEuler Sources directory
