+++ tests/Unit/Evolution/Systems/NewtonianEuler/CMakeLists.txt
@@ -10,7 +10,7 @@ set(LIBRARY_SOURCES
   Test_PrimitiveFromConservative.cpp
   )
 
-add_subdirectory(NewtonianEulerSources)
+add_subdirectory(Sources)
 
 add_test_library(
   ${LIBRARY}
